remove body tag and unnecessary div from components, change name style of components
