+++ project/src/components/movie-page-reviews/movie-page-reviews.tsx
@@ -1,101 +1,6 @@
-function MoviePage(): JSX.Element {
+function MoviePageReviews(): JSX.Element {
   return (
-    <body>
-      <div className="visually-hidden">
-        <svg
-          xmlns="http://www.w3.org/2000/svg"
-          xmlnsXlink="http://www.w3.org/1999/xlink"
-        >
-          <symbol id="add" viewBox="0 0 19 20">
-            <title>+</title>
-            <desc>Created with Sketch.</desc>
-            <g
-              id="Page-1"
-              stroke="none"
-              strokeWidth="1"
-              fill="none"
-              fillRule="evenodd"
-            >
-              <polygon
-                id="+"
-                fill="#EEE5B5"
-                points="10.777832 11.2880859 10.777832 19.5527344 8.41650391 19.5527344 8.41650391 11.2880859 0.627929688 11.2880859 0.627929688 8.92675781 8.41650391 8.92675781 8.41650391 0.662109375 10.777832 0.662109375 10.777832 8.92675781 18.5664062 8.92675781 18.5664062 11.2880859"
-              />
-            </g>
-          </symbol>
-          <symbol id="full-screen" viewBox="0 0 27 27">
-            <path
-              fillRule="evenodd"
-              clipRule="evenodd"
-              d="M23.8571 0H16V3.14286H23.8571V11H27V3.14286V0H23.8571Z"
-              fill="#FFF9D9"
-              fillOpacity="0.7"
-            />
-            <path
-              fillRule="evenodd"
-              clipRule="evenodd"
-              d="M27 23.8571V16H23.8571V23.8571H16V27H23.8571H27L27 23.8571Z"
-              fill="#FFF9D9"
-              fillOpacity="0.7"
-            />
-            <path
-              fillRule="evenodd"
-              clipRule="evenodd"
-              d="M0 3.14286L0 11H3.14286L3.14286 3.14286L11 3.14286V0H3.14286H0L0 3.14286Z"
-              fill="#FFF9D9"
-              fillOpacity="0.7"
-            />
-            <path
-              fillRule="evenodd"
-              clipRule="evenodd"
-              d="M3.14286 27H11V23.8571H3.14286L3.14286 16H0L0 23.8571V27H3.14286Z"
-              fill="#FFF9D9"
-              fillOpacity="0.7"
-            />
-          </symbol>
-          <symbol id="in-list" viewBox="0 0 18 14">
-            <path
-              fillRule="evenodd"
-              clipRule="evenodd"
-              d="M2.40513 5.35353L6.1818 8.90902L15.5807 0L18 2.80485L6.18935 14L0 8.17346L2.40513 5.35353Z"
-              fill="#EEE5B5"
-            />
-          </symbol>
-          <symbol id="pause" viewBox="0 0 14 21">
-            <symbol id="play-s" viewBox="0 0 19 19">
-              <path
-                fillRule="evenodd"
-                clipRule="evenodd"
-                d="M0 0L19 9.5L0 19V0Z"
-                fill="#EEE5B5"
-              />
-            </symbol>
-            <title>Artboard</title>
-            <desc>Created with Sketch.</desc>
-            <g
-              id="Artboard"
-              stroke="none"
-              strokeWidth="1"
-              fill="none"
-              fillRule="evenodd"
-            >
-              <polygon
-                id="Line"
-                fill="#EEE5B5"
-                fillRule="nonzero"
-                points="0 -1.11910481e-13 4 -1.11910481e-13 4 21 0 21"
-              />
-              <polygon
-                id="Line"
-                fill="#EEE5B5"
-                fillRule="nonzero"
-                points="10 -1.11910481e-13 14 -1.11910481e-13 14 21 10 21"
-              />
-            </g>
-          </symbol>
-        </svg>
-      </div>
-
+    <>
       <section className="film-card film-card--full">
         <div className="film-card__hero">
           <div className="film-card__bg">
@@ -142,19 +47,13 @@ function MoviePage(): JSX.Element {
               </p>
 
               <div className="film-card__buttons">
-                <button
-                  className="btn btn--play film-card__button"
-                  type="button"
-                >
+                <button className="btn btn--play film-card__button" type="button">
                   <svg viewBox="0 0 19 19" width="19" height="19">
                     <use xlinkHref="#play-s"></use>
                   </svg>
                   <span>Play</span>
                 </button>
-                <button
-                  className="btn btn--list film-card__button"
-                  type="button"
-                >
+                <button className="btn btn--list film-card__button" type="button">
                   <svg viewBox="0 0 19 20" width="19" height="20">
                     <use xlinkHref="#add"></use>
                   </svg>
@@ -182,7 +81,7 @@ function MoviePage(): JSX.Element {
             <div className="film-card__desc">
               <nav className="film-nav film-card__nav">
                 <ul className="film-nav__list">
-                  <li className="film-nav__item film-nav__item--active">
+                  <li className="film-nav__item">
                     <a href="#" className="film-nav__link">
                       Overview
                     </a>
@@ -192,7 +91,7 @@ function MoviePage(): JSX.Element {
                       Details
                     </a>
                   </li>
-                  <li className="film-nav__item">
+                  <li className="film-nav__item film-nav__item--active">
                     <a href="#" className="film-nav__link">
                       Reviews
                     </a>
@@ -200,40 +99,126 @@ function MoviePage(): JSX.Element {
                 </ul>
               </nav>
 
-              <div className="film-rating">
-                <div className="film-rating__score">8,9</div>
-                <p className="film-rating__meta">
-                  <span className="film-rating__level">Very good</span>
-                  <span className="film-rating__count">240 ratings</span>
-                </p>
-              </div>
+              <div className="film-card__reviews film-card__row">
+                <div className="film-card__reviews-col">
+                  <div className="review">
+                    <blockquote className="review__quote">
+                      <p className="review__text">
+                        Discerning travellers and Wes Anderson fans will
+                        luxuriate in the glorious Mittel-European kitsch of one
+                        of the director&apos;s funniest and most exquisitely designed
+                        films in years.
+                      </p>
 
-              <div className="film-card__text">
-                <p>
-                  In the 1930s, the Grand Budapest Hotel is a popular European
-                  ski resort, presided over by concierge Gustave H. (Ralph
-                  Fiennes). Zero, a junior lobby boy, becomes Gustave&apos;s friend
-                  and protege.
-                </p>
+                      <footer className="review__details">
+                        <cite className="review__author">Kate Muir</cite>
+                        <time className="review__date" dateTime="2016-12-24">
+                          December 24, 2016
+                        </time>
+                      </footer>
+                    </blockquote>
 
-                <p>
-                  Gustave prides himself on providing first-className service to
-                  the hotel&apos;s guests, including satisfying the sexual needs of
-                  the many elderly women who stay there. When one of Gustave&apos;s
-                  lovers dies mysteriously, Gustave finds himself the recipient
-                  of a priceless painting and the chief suspect in her murder.
-                </p>
+                    <div className="review__rating">8,9</div>
+                  </div>
 
-                <p className="film-card__director">
-                  <strong>Director: Wes Anderson</strong>
-                </p>
+                  <div className="review">
+                    <blockquote className="review__quote">
+                      <p className="review__text">
+                        Anderson&apos;s films are too precious for some, but for
+                        those of us willing to lose ourselves in them, they&apos;re a
+                        delight. &quot;The Grand Budapest Hotel&quot; is no different,
+                        except that he has added a hint of gravitas to the mix,
+                        improving the recipe.
+                      </p>
 
-                <p className="film-card__starring">
-                  <strong>
-                    Starring: Bill Murray, Edward Norton, Jude Law, Willem Dafoe
-                    and other
-                  </strong>
-                </p>
+                      <footer className="review__details">
+                        <cite className="review__author">Bill Goodykoontz</cite>
+                        <time className="review__date" dateTime="2015-11-18">
+                          November 18, 2015
+                        </time>
+                      </footer>
+                    </blockquote>
+
+                    <div className="review__rating">8,0</div>
+                  </div>
+
+                  <div className="review">
+                    <blockquote className="review__quote">
+                      <p className="review__text">
+                        I didn&apos;t find it amusing, and while I can appreciate the
+                        creativity, it&apos;s an hour and 40 minutes I wish I could
+                        take back.
+                      </p>
+
+                      <footer className="review__details">
+                        <cite className="review__author">Amanda Greever</cite>
+                        <time className="review__date" dateTime="2015-11-18">
+                          November 18, 2015
+                        </time>
+                      </footer>
+                    </blockquote>
+
+                    <div className="review__rating">8,0</div>
+                  </div>
+                </div>
+                <div className="film-card__reviews-col">
+                  <div className="review">
+                    <blockquote className="review__quote">
+                      <p className="review__text">
+                        The mannered, madcap proceedings are often delightful,
+                        occasionally silly, and here and there, gruesome and/or
+                        heartbreaking.
+                      </p>
+
+                      <footer className="review__details">
+                        <cite className="review__author">Matthew Lickona</cite>
+                        <time className="review__date" dateTime="2016-12-20">
+                          December 20, 2016
+                        </time>
+                      </footer>
+                    </blockquote>
+
+                    <div className="review__rating">7,2</div>
+                  </div>
+
+                  <div className="review">
+                    <blockquote className="review__quote">
+                      <p className="review__text">
+                        It is certainly a magical and childlike way of
+                        storytelling, even if the content is a little more
+                        adult.
+                      </p>
+
+                      <footer className="review__details">
+                        <cite className="review__author">Paula Fleri-Soler</cite>
+                        <time className="review__date" dateTime="2016-12-20">
+                          December 20, 2016
+                        </time>
+                      </footer>
+                    </blockquote>
+
+                    <div className="review__rating">7,6</div>
+                  </div>
+
+                  <div className="review">
+                    <blockquote className="review__quote">
+                      <p className="review__text">
+                        It is certainly a magical and childlike way of
+                        storytelling, even if the content is a little more
+                        adult.
+                      </p>
+
+                      <footer className="review__details">
+                        <cite className="review__author">Paula Fleri-Soler</cite>
+                        <time className="review__date" dateTime="2016-12-20">
+                          December 20, 2016
+                        </time>
+                      </footer>
+                    </blockquote>
+
+                    <div className="review__rating">7,0</div>
+                  </div>
+                </div>
               </div>
             </div>
           </div>
@@ -325,8 +310,8 @@ function MoviePage(): JSX.Element {
           </div>
         </footer>
       </div>
-    </body>
+    </>
   );
 }
 
-export default MoviePage;
+export default MoviePageReviews;
